Candidate profile:expectedCtc is not coming in response. #731

+++ server/src/main/java/io/litmusblox/server/model/JobCandidateMapping.java
@@ -159,7 +159,7 @@ public class JobCandidateMapping implements Serializable {
     private String candidateRejectionValue;
 
     @Column(name = "EXPECTED_CTC")
-    private Long expectedCtc;
+    private Double expectedCtc;
 
     @Column(name = "PERCENTAGE_HIKE")
     private Long percentageHike;
